Added FTP ability
Added the ability to submit files to an ftp server. It doesn't support TLS or SSL so it has to be send in the clear.

Nature of the script is to do basic script, and at the end, if its set to allow FTP it will upload the files and then delete them from local when they are uploaded